Extract getLocalAppId to command class
